src/kafka_client/requirements.txt
kafka-python>=2.0.2
jsonschema>=4.17.3
pytest>=7.3.1
python-dotenv>=0.19.0
cryptography>=3.4.7  # Required for SSL support 